requirements.txt
Flask>=2.0
Authlib>=1.0
python-dotenv>=0.15
requests>=2.32